- Renamed owntpm package to tpmmgr - addressed comments from Erik
